+++ articoli/tests.py
@@ -3,10 +3,12 @@ import random
 import string
 
 from django.core.files.temp import NamedTemporaryFile
+from django.core.urlresolvers import reverse
 from django.test import TestCase
 
 from anagrafica.models import Delega
 from anagrafica.permessi.applicazioni import PRESIDENTE
+from autenticazione.utils_test import TestFunzionale
 from base.files import Zip
 from base.utils_tests import crea_persona, crea_persona_sede_appartenenza
 
@@ -39,13 +41,11 @@ class ArticoliTests(TestCase):
         )
         delega_presidente_in_corso.save()
 
-
         articolo = Articolo.objects.create(
                 titolo='Titolo 1',
                 corpo=parola_casuale(3000),
                 data_inizio_pubblicazione=datetime.datetime.now() - datetime.timedelta(days=5),
             )
-
 
         self.assertEqual(articolo.corpo[:Articolo.DIMENSIONE_ESTRATTO], articolo.estratto)
         self.assertFalse(articolo.termina)
@@ -80,7 +80,6 @@ class ArticoliTests(TestCase):
             segmento='B',
             articolo=articolo3
         )
-
 
         z = Zip(oggetto=articolo3)
         f1 = NamedTemporaryFile(delete=False, mode='wt')
@@ -123,3 +122,158 @@ class ArticoliTests(TestCase):
         articoli_presidente = segmenti_presidente.oggetti_collegati()
         self.assertIn(articolo2, articoli_presidente)
         self.assertIn(articolo3, articoli_presidente)
+
+
+class TestFunzionaleArticoli(TestFunzionale):
+
+    def test_lista_articoli_vuota(self):
+        persona = crea_persona()
+        persona, sede, app = crea_persona_sede_appartenenza()
+        sessione_persona = self.sessione_utente(persona=persona)
+        sessione_persona.visit("%s%s" % (self.live_server_url, reverse('lista_articoli')))
+        self.assertTrue(sessione_persona.is_text_present('Non è stato trovato alcun articolo'))
+
+    def test_lista_articoli(self):
+        articolo = Articolo.objects.create(
+            titolo='Titolo 1981',
+            corpo='Testo random',
+            estratto='qualcosa',
+            data_inizio_pubblicazione='1981-12-10',
+            stato=Articolo.PUBBLICATO
+        )
+        articolo2 = Articolo.objects.create(
+            titolo='Titolo primo 1980',
+            corpo='Testo random 2',
+            estratto='un pezzo',
+            data_inizio_pubblicazione='1980-06-10',
+            stato=Articolo.PUBBLICATO
+        )
+        articolo3 = Articolo.objects.create(
+            titolo='Titolo secondo 1980',
+            corpo='Testo random 3',
+            estratto='una parte',
+            data_inizio_pubblicazione='1980-12-10',
+            stato=Articolo.PUBBLICATO
+        )
+        persona = crea_persona()
+        persona, sede, app = crea_persona_sede_appartenenza()
+        sessione_persona = self.sessione_utente(persona=persona)
+        sessione_persona.visit("%s%s" % (self.live_server_url, reverse('lista_articoli')))
+        self.assertFalse(sessione_persona.is_text_present('Non è stato trovato alcun articolo'))
+        self.assertTrue(sessione_persona.is_text_present(articolo.titolo))
+        self.assertTrue(sessione_persona.is_text_present(articolo.estratto))
+        self.assertTrue(sessione_persona.is_text_present(articolo2.titolo))
+        self.assertTrue(sessione_persona.is_text_present(articolo2.estratto))
+        self.assertTrue(sessione_persona.is_text_present(articolo3.titolo))
+        self.assertTrue(sessione_persona.is_text_present(articolo3.estratto))
+        self.assertEqual(3, len(sessione_persona.find_by_css('.panel.panel-primary')))
+        sessione_persona.fill('q', 'primo')
+        sessione_persona.find_by_xpath('//button[@type="submit"]').first.click()
+        self.assertEqual(1, len(sessione_persona.find_by_css('.panel.panel-primary')))
+        self.assertFalse(sessione_persona.is_text_present(articolo.titolo))
+        self.assertFalse(sessione_persona.is_text_present(articolo.estratto))
+        self.assertTrue(sessione_persona.is_text_present(articolo2.titolo))
+        self.assertTrue(sessione_persona.is_text_present(articolo2.estratto))
+        self.assertFalse(sessione_persona.is_text_present(articolo3.titolo))
+        self.assertFalse(sessione_persona.is_text_present(articolo3.estratto))
+        sessione_persona.find_by_xpath('//select[@name="anno"]//option[@value="1980"]').first.click()
+        sessione_persona.find_by_xpath('//button[@type="submit"]').first.click()
+        self.assertEqual(1, len(sessione_persona.find_by_css('.panel.panel-primary')))
+        self.assertFalse(sessione_persona.is_text_present(articolo.titolo))
+        self.assertFalse(sessione_persona.is_text_present(articolo.estratto))
+        self.assertTrue(sessione_persona.is_text_present(articolo2.titolo))
+        self.assertTrue(sessione_persona.is_text_present(articolo2.estratto))
+        self.assertFalse(sessione_persona.is_text_present(articolo3.titolo))
+        self.assertFalse(sessione_persona.is_text_present(articolo3.estratto))
+        sessione_persona.fill('q', '')
+        sessione_persona.find_by_xpath('//select[@name="anno"]//option[@value="1980"]').first.click()
+        sessione_persona.find_by_xpath('//button[@type="submit"]').first.click()
+        self.assertEqual(2, len(sessione_persona.find_by_css('.panel.panel-primary')))
+        self.assertFalse(sessione_persona.is_text_present(articolo.titolo))
+        self.assertFalse(sessione_persona.is_text_present(articolo.estratto))
+        self.assertTrue(sessione_persona.is_text_present(articolo2.titolo))
+        self.assertTrue(sessione_persona.is_text_present(articolo2.estratto))
+        self.assertTrue(sessione_persona.is_text_present(articolo3.titolo))
+        self.assertTrue(sessione_persona.is_text_present(articolo3.estratto))
+        sessione_persona.find_by_xpath('//select[@name="anno"]//option[@value="1980"]').first.click()
+        sessione_persona.find_by_xpath('//select[@name="mese"]//option[@value=6]').first.click()
+        sessione_persona.find_by_xpath('//button[@type="submit"]').first.click()
+        self.assertEqual(1, len(sessione_persona.find_by_css('.panel.panel-primary')))
+        self.assertFalse(sessione_persona.is_text_present(articolo.titolo))
+        self.assertFalse(sessione_persona.is_text_present(articolo.estratto))
+        self.assertTrue(sessione_persona.is_text_present(articolo2.titolo))
+        self.assertTrue(sessione_persona.is_text_present(articolo2.estratto))
+        self.assertFalse(sessione_persona.is_text_present(articolo3.titolo))
+        self.assertFalse(sessione_persona.is_text_present(articolo3.estratto))
+        sessione_persona.fill('q', '')
+        sessione_persona.find_by_xpath('//button[@type="submit"]').first.click()
+        sessione_persona.find_link_by_partial_text('Leggi').first.click()
+        self.assertTrue(sessione_persona.is_text_present(articolo.titolo))
+        self.assertTrue(sessione_persona.is_text_present(articolo.corpo))
+        self.assertFalse(sessione_persona.is_text_present(articolo2.titolo))
+        self.assertFalse(sessione_persona.is_text_present(articolo2.corpo))
+        self.assertFalse(sessione_persona.is_text_present(articolo3.titolo))
+        self.assertFalse(sessione_persona.is_text_present(articolo3.corpo))
+        self.assertTrue(sessione_persona.is_text_present('Letture'))
+
+    def test_dettaglio_articoli_pubblici(self):
+        articolo = Articolo.objects.create(
+            titolo='Titolo 1981',
+            corpo='Testo random',
+            estratto='qualcosa',
+            data_inizio_pubblicazione='1981-12-10',
+            stato=Articolo.PUBBLICATO
+        )
+        articolo2 = Articolo.objects.create(
+            titolo='Titolo primo 1980',
+            corpo='Testo random 2',
+            estratto='un pezzo',
+            data_inizio_pubblicazione='1980-06-10',
+            stato=Articolo.PUBBLICATO
+        )
+        segmento_presidenti_no_filtri = ArticoloSegmento.objects.create(
+            segmento='I',
+            articolo=articolo2,
+        )
+        sessione_persona = self.sessione_anonimo()
+        sessione_persona.visit("%s%s" % (
+            self.live_server_url, reverse('dettaglio_articolo', kwargs={
+                'articolo_slug': articolo.slug
+            })
+        ))
+        self.assertTrue(sessione_persona.is_text_present(articolo.titolo))
+
+        sessione_persona.visit("%s%s" % (
+            self.live_server_url, reverse('dettaglio_articolo', kwargs={
+                'articolo_slug': articolo2.slug
+            })
+        ))
+        self.assertFalse(sessione_persona.is_text_present(articolo2.titolo))
+        self.assertTrue(sessione_persona.is_text_present('404'))
+
+    def test_dettaglio_articoli_privati(self):
+        articolo = Articolo.objects.create(
+            titolo='Titolo 1981',
+            corpo='Testo random',
+            estratto='qualcosa',
+            data_inizio_pubblicazione='1981-12-10',
+            stato=Articolo.PUBBLICATO
+        )
+        presidente = crea_persona()
+        presidente.save()
+        presidente, sede, _ = crea_persona_sede_appartenenza(presidente)
+        delega_presidente_in_corso = Delega(
+            persona=presidente,
+            tipo=PRESIDENTE,
+            oggetto=sede,
+            inizio=datetime.datetime.now() - datetime.timedelta(days=5),
+            fine=datetime.datetime.now() + datetime.timedelta(days=5)
+        )
+        delega_presidente_in_corso.save()
+        sessione_persona = self.sessione_utente(persona=presidente)
+        sessione_persona.visit("%s%s" % (
+            self.live_server_url, reverse('dettaglio_articolo', kwargs={
+                'articolo_slug': articolo.slug
+            })
+        ))
+        self.assertTrue(sessione_persona.is_text_present(articolo.titolo))
