+++ gestione_file/tests.py
@@ -5,10 +5,12 @@ import tempfile
 import django.core.files
 from django.core.urlresolvers import reverse
 from django.test import TestCase
+
+from autenticazione.utils_test import TestFunzionale
 from filer.models import Folder
 from filer.tests import create_image
 
-from base.utils_tests import crea_persona_sede_appartenenza, crea_sede
+from base.utils_tests import crea_persona_sede_appartenenza, crea_sede, crea_persona
 from curriculum.models import Titolo, TitoloPersonale
 from gestione_file.models import Documento, DocumentoSegmento, Immagine
 
@@ -177,3 +179,94 @@ class TestSegmenti(TestCase):
         oggetti = documenti.oggetti_collegati()
         self.assertEqual(oggetti.count(), 1)
         self.assertEqual(set(oggetti), set(Documento.objects.filter(pk=self.file_2.pk)))
+
+
+class TestFunzionaleGestioneFile(TestFunzionale):
+
+    def setUp(self):
+
+        from django.contrib.auth import get_user_model
+        User = get_user_model()
+        try:
+            User.objects.get(email='admin@admin.it')
+        except User.DoesNotExist:
+            self.superuser = User.objects.create_superuser(
+                'admin@admin.it', 'secret'
+            )
+        self.client.login(username='admin@admin.it', password='secret')
+        self.img = create_image()
+        file_obj, self.filename = tempfile.mkstemp('.jpg')
+        self.image_name = os.path.basename(self.filename)
+        self.img.save(self.filename, 'JPEG')
+        file_obj, self.doc_filename = tempfile.mkstemp('.doc')
+        self.doc_name = os.path.basename(self.doc_filename)
+        with open(self.doc_filename, 'w') as file_obj:
+            file_obj.write('text')
+        super(TestFunzionaleGestioneFile, self).setUp()
+
+    def tearDown(self):
+        self.client.logout()
+        os.remove(self.filename)
+        os.remove(self.doc_filename)
+        super(TestFunzionaleGestioneFile, self).tearDown()
+
+    def test_lista_documenti_vuota(self):
+        persona = crea_persona()
+        persona, sede, app = crea_persona_sede_appartenenza()
+        sessione_persona = self.sessione_utente(persona=persona)
+        sessione_persona.visit("%s%s" % (self.live_server_url,
+                                            reverse('lista_documenti')))
+        self.assertTrue(sessione_persona.is_text_present('tipo'))
+        self.assertTrue(sessione_persona.is_text_present('nome'))
+        self.assertTrue(sessione_persona.is_text_present('dimensione'))
+        self.assertTrue(sessione_persona.is_text_present('accessi'))
+        self.assertEqual(0, len(sessione_persona.find_by_tag('tbody').find_by_tag('tr')))
+
+    def test_lista_documenti(self):
+        extra_headers = {}
+        self.assertEqual(Documento.objects.count(), 0)
+        folder_radice = Folder.objects.create(name='radice')
+        folder_figlia = Folder.objects.create(name='sottocartella', parent=folder_radice)
+        file_obj = django.core.files.File(open(self.doc_filename, 'rb'))
+        url = reverse('admin:filer-ajax_upload', kwargs={'folder_id': folder_figlia.pk})
+        post_data = {
+            'Filename': self.doc_name,
+            'Filedata': file_obj,
+            'jsessionid': self.client.session.session_key
+        }
+        self.client.post(url, post_data, **extra_headers)
+        self.assertEqual(Documento.objects.count(), 1)
+        uploaded_file = Documento.objects.all()[0]
+        self.assertEqual(uploaded_file.original_filename, self.doc_name)
+
+        url = reverse('admin:gestione_file_documento_add') + '?parent_id={}'.format(folder_radice.pk)
+        post_data = {
+            'url_documento': 'http://www.example.com',
+            'segmenti-TOTAL_FORMS': 0,
+            'segmenti-INITIAL_FORMS': 0,
+            'segmenti-MIN_NUM_FORMS': 0,
+            'segmenti-MAX_NUM_FORMS': 0
+        }
+        self.client.post(url, post_data, **extra_headers)
+        self.assertEqual(Documento.objects.count(), 2)
+        uploaded_file = Documento.objects.all()[1]
+        self.assertEqual(uploaded_file.folder, folder_radice)
+        self.assertEqual(uploaded_file.original_filename, 'www.example.com')
+
+        persona = crea_persona()
+        persona, sede, app = crea_persona_sede_appartenenza()
+        sessione_persona = self.sessione_utente(persona=persona)
+        sessione_persona.visit("%s%s" % (self.live_server_url,
+                                            reverse('lista_documenti')))
+        self.assertTrue(sessione_persona.is_text_present('tipo'))
+        self.assertTrue(sessione_persona.is_text_present('nome'))
+        self.assertTrue(sessione_persona.is_text_present('dimensione'))
+        self.assertTrue(sessione_persona.is_text_present('accessi'))
+        self.assertEqual(1, len(sessione_persona.find_by_tag('tbody').find_by_tag('tr')))
+        sessione_persona.find_link_by_text('radice').first.click()
+        self.assertTrue(sessione_persona.is_text_present('www.example.com'))
+        sessione_persona.find_link_by_text('sottocartella').first.click()
+        self.assertTrue(sessione_persona.is_text_present(self.doc_name))
+        sessione_persona.fill('q', 'example')
+        sessione_persona.find_by_xpath('//button[@type="submit"]').first.click()
+        self.assertTrue(sessione_persona.is_text_present('www.example.com'))
